fix: full codebase cleanup — clippy warnings, panics, dead code, type consistency
- Remove all #[allow(dead_code)] and #[allow(clippy::...)] suppressions
- Fix 5 clippy warnings (&Box<T>, unwrap after is_some, collapsible_if, manual_range_contains)
- Replace all unwrap()/expect() with graceful error handling
- Unify float types: Current fields f32 → f64 (matches Daily)
- Extract calc_and_fetch params instead of re-parsing Args::parse()
- Deduplicate depictors match arms, fix hardcoded chart range
- Fix cache underflow with checked_sub, remove dead code and stale comments

+++ src/response_layouts.rs
@@ -1,6 +1,5 @@
 use serde::{Deserialize, Serialize};
 
-#[allow(dead_code)]
 #[derive(Debug, Deserialize, Serialize, Clone)]
 pub struct WeatherDaily {
     pub latitude: f64,
@@ -13,7 +12,7 @@ pub struct WeatherDaily {
     pub daily_units: DailyUnits,
     pub daily: Daily,
 }
-#[allow(dead_code)]
+
 #[derive(Debug, Deserialize, Serialize, Clone)]
 pub struct WeatherResponse {
     pub latitude: f64,
@@ -25,7 +24,7 @@ pub struct WeatherResponse {
     pub current_units: CurrentUnits,
     pub current: Current,
 }
-#[allow(dead_code)]
+
 #[derive(Debug, Deserialize, Serialize, Clone)]
 pub struct DailyUnits {
     pub time: String,
@@ -37,7 +36,7 @@ pub struct DailyUnits {
     pub temperature_2m_mean: String,
     pub wind_speed_10m_max: String,
 }
-#[allow(dead_code)]
+
 #[derive(Debug, Deserialize, Serialize, Clone)]
 pub struct Daily {
     pub time: Vec<String>,
@@ -49,7 +48,7 @@ pub struct Daily {
     pub temperature_2m_mean: Vec<f64>,
     pub wind_speed_10m_max: Vec<f64>,
 }
-#[allow(dead_code)]
+
 #[derive(Debug, Deserialize, Serialize, Clone)]
 pub struct CurrentUnits {
     pub temperature_2m: String,
@@ -58,15 +57,15 @@ pub struct CurrentUnits {
     pub snowfall: String,
     pub precipitation: String,
 }
-#[allow(dead_code)]
+
 #[derive(Debug, Deserialize, Serialize, Clone)]
 pub struct Current {
     pub time: String,
-    pub temperature_2m: f32,
-    pub wind_speed_10m: f32,
-    pub rain: f32,
-    pub snowfall: f32,
-    pub precipitation: f32,
+    pub temperature_2m: f64,
+    pub wind_speed_10m: f64,
+    pub rain: f64,
+    pub snowfall: f64,
+    pub precipitation: f64,
 }
 
 #[derive(Debug, Deserialize, Serialize, Clone)]
@@ -74,9 +73,3 @@ pub struct Location {
     pub lat: f64,
     pub lon: f64,
 }
-
-// #[allow(dead_code)]
-// #[derive(Debug, Deserialize)]
-// pub struct IpLanLot {
-//
-// }
